+++ MantidPlot/CMakeLists.txt
@@ -131,7 +131,6 @@ set ( QTIPLOT_SRCS src/ApplicationWindow.cpp
                    src/TiledWindow.cpp
                    src/TitlePicker.cpp
                    src/TranslateCurveTool.cpp
-                   src/TSVSerialiser.cpp
                    src/UserFunction.cpp
                    src/VectorCurve.cpp
                    src/WindowFactory.cpp
@@ -333,7 +332,6 @@ set ( QTIPLOT_HDRS src/ApplicationWindow.h
                    src/TiledWindow.h
                    src/TitlePicker.h
                    src/TranslateCurveTool.h
-                   src/TSVSerialiser.h
                    src/UserFunction.h
                    src/VectorCurve.h
                    src/WindowFactory.h
