+++ MantidPlot/src/Grid.cpp
@@ -31,7 +31,7 @@
 #include "Grid.h"
 #include "ColorBox.h"
 
-#include "TSVSerialiser.h"
+#include "MantidQtAPI/TSVSerialiser.h"
 
 #include <qwt_plot_canvas.h>
 #include <qwt_painter.h>
