+++ MantidPlot/src/Mantid/MantidMatrix.cpp
@@ -10,7 +10,7 @@
 #include "Preferences.h"
 #include "../pixmaps.h"
 
-#include "TSVSerialiser.h"
+#include "MantidQtAPI/TSVSerialiser.h"
 
 #include "MantidAPI/MatrixWorkspace.h"
 #include "MantidAPI/NumericAxis.h"
